// TestAngular/src/app/models/Pessoa.ts
export class Pessoa {

  constructor() {
    this.id = 0;
    this.cpf = 0;
    this.nome = '';
    this.sobrenome = '';
    this.nacionalidade = '';
    this.cep = '';
    this.estado = '';
    this.cidade = '';
    this.logradouro = '';
    this.email = '';
    this.telefone = '';
  }

  id: number;
  cpf: number;
  nome: string;
  sobrenome: string;
  nacionalidade: string;
  cep: string;
  estado: string;
  cidade: string;
  logradouro: string;
  email: string;
  telefone: string;
}